feat: Add new nft page
