+++ package.json
@@ -29,7 +29,7 @@
     "@material-ui/icons": "^4.11.2",
     "@material-ui/lab": "^4.0.0-alpha.57",
     "@metamask/detect-provider": "^1.2.0",
-    "@mibao-ui/components": "0.0.8",
+    "@mibao-ui/components": "0.0.10",
     "@nervosnetwork/ckb-sdk-core": "^0.40.0",
     "@sentry/react": "^6.9.0",
     "@sentry/tracing": "^6.9.0",
